serialization tests cleanup
