+++ Tests/SwiftCraftTests/SerializationTests/NumberSerializationTests.swift
@@ -1,8 +1,8 @@
 //
-//  CodableDataTypesTest.swift
-//  SwiftCraftTests
+//  NumberSerializationTests.swift
+//  SwiftCraft iOS
 //
-//  Created by Noah Peeters on 21.05.18.
+//  Created by Noah Peeters on 25.06.18.
 //  Copyright © 2018 Noah Peeters. All rights reserved.
 //
 
@@ -10,7 +10,7 @@ import Quick
 import Nimble
 @testable import SwiftCraft
 
-public class BoolCodableTests: QuickSpec {
+public class BoolSerializationTests: QuickSpec {
     public override func spec() {
         describe("serializing") {
             it("serializes false") {
@@ -38,7 +38,7 @@ public class BoolCodableTests: QuickSpec {
     }
 }
 
-public class UInt8CodableTest: QuickSpec {
+public class UInt8SerializationTests: QuickSpec {
     public override func spec() {
         describe("serializing") {
             it("serializes correctly") {
@@ -58,7 +58,7 @@ public class UInt8CodableTest: QuickSpec {
     }
 }
 
-public class Int8CodableTest: QuickSpec {
+public class Int8SerializationTests: QuickSpec {
     public override func spec() {
         describe("serializing") {
             it("serializes positive number correctly") {
@@ -85,7 +85,7 @@ public class Int8CodableTest: QuickSpec {
     }
 }
 
-public class UInt64CodableTest: QuickSpec {
+public class UInt64SerializationTests: QuickSpec {
     public override func spec() {
         describe("serializing") {
             it("serializes correctly") {
@@ -109,7 +109,7 @@ public class UInt64CodableTest: QuickSpec {
     }
 }
 
-public class FloatCodableTest: QuickSpec {
+public class FloatSerializationTests: QuickSpec {
     public override func spec() {
         let serializedData: ByteArray = [0b01000001, 0b10111000, 0, 0]
         let deserializedFloat: Float = 23
@@ -136,7 +136,7 @@ public class FloatCodableTest: QuickSpec {
     }
 }
 
-public class DoubleCodableTest: QuickSpec {
+public class DoubleSerializationTests: QuickSpec {
     public override func spec() {
         let serializedData: ByteArray = [0b01000000, 0b00110111, 0, 0, 0, 0, 0, 0]
         let deserializedDouble: Double = 23
@@ -163,7 +163,7 @@ public class DoubleCodableTest: QuickSpec {
     }
 }
 
-public class VarInt32CodableTest: QuickSpec {
+public class VarInt32SerializationTests: QuickSpec {
     public override func spec() {
         let data: [Int32: ByteArray] = [
             0: [0x00],
@@ -218,7 +218,7 @@ public class VarInt32CodableTest: QuickSpec {
     }
 }
 
-public class VarInt64CodableTest: QuickSpec {
+public class VarInt64SerializationTests: QuickSpec {
     public override func spec() {
         let data: [Int64: ByteArray] = [
             0: [0x00],
@@ -270,147 +270,6 @@ public class VarInt64CodableTest: QuickSpec {
                     let invalidData: ByteArray = [0xff, 0xff, 0xff, 0xff]
                     expect { try VarInt64(from: invalidData) }.to(throwError(BufferError.noDataAvailable))
                 }
-            }
-        }
-    }
-}
-
-public class StringCodableTest: QuickSpec {
-    public override func spec() {
-        describe("ascii") {
-            let serializedData: ByteArray = [12, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x20, 0x57, 0x6f, 0x72, 0x6c, 0x64, 0x21]
-            let deserializedStrign = "Hello World!"
-
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect(deserializedStrign.directSerialized()).to(equal(serializedData))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? String(from: serializedData)).to(equal(deserializedStrign))
-                }
-            }
-        }
-
-        describe("emoji") {
-            let serializedData: ByteArray = [4, 0xf0, 0x9f, 0x98, 0x8e]
-            let deserializedString = "😎"
-
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect(deserializedString.directSerialized()).to(equal(serializedData))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? String(from: serializedData)).to(equal(deserializedString))
-                }
-            }
-        }
-
-        describe("empty string") {
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect("".directSerialized()).to(equal([0]))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? String(from: [0])).to(equal(""))
-                }
-            }
-        }
-
-        describe("deserializing invalid data") {
-            it("throws when deserializing empty data") {
-                expect { try String(from: []) }.to(throwError(BufferError.noDataAvailable))
-            }
-
-            it("throws when length is longer then the remaining data") {
-                expect { try String(from: [5, 42, 42]) }.to(throwError(BufferError.noDataAvailable))
-            }
-
-            it("throws when the data is not valid utf8") {
-                expect {
-                    try String(from: [3, 0xf0, 0x9f, 0x98])
-                }.to(throwError(String.TypeDeserializeError.invalidStringData))
-            }
-        }
-    }
-}
-
-public class PositionCodableTest: QuickSpec {
-    public override func spec() {
-        describe("zero position") {
-            let serializedData: ByteArray = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
-            let deserializedPosition = BlockPosition(x: 0, y: 0, z: 0)
-
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect(deserializedPosition.directSerialized()).to(equal(serializedData))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? BlockPosition(from: serializedData)).to(equal(deserializedPosition))
-                }
-            }
-        }
-
-        describe("positive position") {
-            let serializedData: ByteArray = [0x00, 0x00, 0x0a, 0x80, 0xac, 0x00, 0x00, 0x2c]
-            let deserializedPosition = BlockPosition(x: 42, y: 43, z: 44)
-
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect(deserializedPosition.directSerialized()).to(equal(serializedData))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? BlockPosition(from: serializedData)).to(equal(deserializedPosition))
-                }
-            }
-        }
-
-        describe("negative position") {
-            let serializedData: ByteArray = [0xff, 0xff, 0xf5, 0x80, 0x03, 0xff, 0xff, 0xd4]
-            let deserializedPosition = BlockPosition(x: -42, y: 0, z: -44)
-
-            describe("serializing") {
-                it("serializes correctly") {
-                    expect(deserializedPosition.directSerialized()).to(equal(serializedData))
-                }
-            }
-
-            describe("deserializing") {
-                it("deserializes correctly") {
-                    expect(try? BlockPosition(from: serializedData)).to(equal(deserializedPosition))
-                }
-            }
-        }
-    }
-}
-
-public class UUIDCodableTest: QuickSpec {
-    public override func spec() {
-        describe("when creating a uuid") {
-            let uuid = UUID()
-
-            it("serialized uuid is 16 bytes long") {
-                expect(uuid.directSerialized().count).to(equal(16))
-            }
-
-            it("can be serialized and deserialized") {
-                let serializedUUID = uuid.directSerialized()
-                let deserializedUUID = try? UUID(from: serializedUUID)
-                expect(deserializedUUID).to(equal(uuid))
             }
         }
     }
